# storage/api-docs/api-docs-admin-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API",
        "description": "API documentation for the School Management System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/"
        }
    ],
    "paths": {
...
        "/admin/schedule": {
            "post": {
                "tags": [
                    "Schedule Management"
                ],
                "summary": "Create a New Schedule",
                "description": "Add a new schedule with specific days and time slots.",
                "operationId": "aba28ed83715b0d3d324f48a918efa1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "days",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "days": {
                                        "description": "Days of the schedule",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "Monday",
                                                "Tuesday",
                                                "Wednesday",
                                                "Thursday",
                                                "Friday",
                                                "Saturday",
                                                "Sunday"
                                            ],
                                            "example": "Monday"
                                        }
                                    },
                                    "start_time": {
                                        "description": "Start time of the schedule",
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00:00"
                                    },
                                    "end_time": {
                                        "description": "End time of the schedule (must be after start_time)",
                                        "type": "string",
                                        "format": "time",
                                        "example": "16:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Schedule created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Schedule created successfully"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/schedule/{id}": {
            "put": {
                "tags": [
                    "Schedule Management"
                ],
                "summary": "Update an Existing Schedule",
                "description": "Update the details of an existing schedule.",
                "operationId": "700e6b638844d3590535136188fe760e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the schedule to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "days",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "days": {
                                        "description": "Days of the schedule",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "Monday",
                                                "Tuesday",
                                                "Wednesday",
                                                "Thursday",
                                                "Friday",
                                                "Saturday",
                                                "Sunday"
                                            ],
                                            "example": "Tuesday"
                                        }
                                    },
                                    "start_time": {
                                        "description": "Start time of the schedule",
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00:00"
                                    },
                                    "end_time": {
                                        "description": "End time of the schedule (must be after start_time)",
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Schedule updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Schedule updated successfully"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Schedule not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Schedule Management"
                ],
                "summary": "Delete a Schedule",
                "description": "Delete an existing schedule by its ID.",
                "operationId": "b2e4e462eed615bd49696969ee8d0b37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the schedule to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Schedule deleted successfully"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Schedule not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}